+++ YaksRPG/Models/ICharacterClass.cs
@@ -8,7 +8,7 @@ public interface ICharacterClass
   
   public DiceRoll HitDicePerLevel { get; }
 
-  public int AttackBonusPerLevel { get; }
+  public float AttackBonusPerLevel { get; }
   
   public IEnumerable<Feature> Features { get; }
 }
